Move Stats store into store folder.
